Reorganize CommitMetric location, satisfy linter
